Exception handling for coping strategy and safetyplan services.

+++ safetyplan/src/main/java/com/cove/safetyplan/service/CopingStrategyServiceImpl.java
@@ -1,8 +1,13 @@
 package com.cove.safetyplan.service;
 
 import com.cove.safetyplan.dto.model.CopingStrategyDto;
+import com.cove.safetyplan.exception.CoveException;
+import com.cove.safetyplan.exception.EntityType;
+import com.cove.safetyplan.exception.ExceptionType;
 import com.cove.safetyplan.model.entities.CopingStrategy;
+import com.cove.safetyplan.model.entities.SafetyPlan;
 import com.cove.safetyplan.repository.JPACopingStrategyRepository;
+import com.cove.safetyplan.repository.JPASafetyPlanRepository;
 import com.netflix.discovery.EurekaClient;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +17,8 @@ import javax.swing.text.html.Option;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import static com.cove.safetyplan.exception.EntityType.*;
+import static com.cove.safetyplan.exception.ExceptionType.*;
 
 @Component
 public class CopingStrategyServiceImpl implements CopingStrategyService {
@@ -25,11 +32,18 @@ public class CopingStrategyServiceImpl implements CopingStrategyService {
     @Autowired
     private JPACopingStrategyRepository jpaCopingStrategyRepository;
 
+    @Autowired
+    private JPASafetyPlanRepository jpaSafetyPlanRepository;
+
     @Override
     public List<CopingStrategyDto> addCopingStrategyToSafetyPlan(CopingStrategyDto copingStrategyDto){
-        CopingStrategy newCopingStrategy = modelMapper.map(copingStrategyDto, CopingStrategy.class);
-        jpaCopingStrategyRepository.save(newCopingStrategy);
-        return getCopingStrategies(copingStrategyDto.getSafetyplanId());
+        Optional<SafetyPlan> safetyPlan = jpaSafetyPlanRepository.findById(copingStrategyDto.getSafetyplanId());
+        if(safetyPlan.isPresent()){
+            CopingStrategy newCopingStrategy = modelMapper.map(copingStrategyDto, CopingStrategy.class);
+            jpaCopingStrategyRepository.save(newCopingStrategy);
+            return getCopingStrategies(copingStrategyDto.getSafetyplanId());
+        }
+        throw exception(SAFETYPLAN, ENTITY_NOT_FOUND, Long.toString(copingStrategyDto.getSafetyplanId()));
     }
 
     @Override
@@ -40,20 +54,28 @@ public class CopingStrategyServiceImpl implements CopingStrategyService {
             copingStrategyModel.setType(copingStrategyDto.getType());
             return modelMapper.map(jpaCopingStrategyRepository.save(copingStrategyModel), CopingStrategyDto.class);
         }
-        return new CopingStrategyDto();
+        throw exception(COPINGSTRATEGY, ENTITY_NOT_FOUND, Long.toString(copingStrategyDto.getCopingStrategyId()));
     }
 
     @Override
-    public List<CopingStrategyDto> getCopingStrategies(long safetyPlanId){
-        List<CopingStrategyDto> copingStrategyDtos = new ArrayList<>();
-        jpaCopingStrategyRepository.findBySafetyplanId(safetyPlanId)
-                .forEach(strategy -> copingStrategyDtos.add(modelMapper.map(strategy, CopingStrategyDto.class)));
-        return copingStrategyDtos;
+    public List<CopingStrategyDto> getCopingStrategies(long safetyplanId){
+        Optional<SafetyPlan> safetyPlan = jpaSafetyPlanRepository.findById(safetyplanId);
+        if(safetyPlan.isPresent()){
+            List<CopingStrategyDto> copingStrategyDtos = new ArrayList<>();
+            jpaCopingStrategyRepository.findBySafetyplanId(safetyplanId)
+                    .forEach(strategy -> copingStrategyDtos.add(modelMapper.map(strategy, CopingStrategyDto.class)));
+            return copingStrategyDtos;
+        }
+        throw exception(SAFETYPLAN, ENTITY_NOT_FOUND, Long.toString(safetyplanId));
     }
 
     @Override
     public CopingStrategyDto getCopingStrategy(long copingStrategyId){
-        return modelMapper.map(jpaCopingStrategyRepository.findById(copingStrategyId).get(), CopingStrategyDto.class);
+        Optional<CopingStrategy> copingStrategy = jpaCopingStrategyRepository.findById(copingStrategyId);
+        if(copingStrategy.isPresent()){
+            return modelMapper.map(copingStrategy.get(), CopingStrategyDto.class);
+        }
+        throw exception(COPINGSTRATEGY, ENTITY_NOT_FOUND, Long.toString(copingStrategyId));
     }
 
     @Override
@@ -62,13 +84,22 @@ public class CopingStrategyServiceImpl implements CopingStrategyService {
         if (copingStrategy.isPresent()){
             jpaCopingStrategyRepository.delete(copingStrategy.get());
         }
+        throw exception(COPINGSTRATEGY, ENTITY_NOT_FOUND, Long.toString(copingStrategyId));
     }
 
     @Override
     public void deleteAllCopingStrategiesFromSafetyPlan(long safetyplanId){
-        List<CopingStrategy> copingStrategies = jpaCopingStrategyRepository.findBySafetyplanId(safetyplanId);
+        Optional<SafetyPlan> safetyPlan = jpaSafetyPlanRepository.findById(safetyplanId);
+        if(safetyPlan.isPresent()){
+            List<CopingStrategy> copingStrategies = jpaCopingStrategyRepository.findBySafetyplanId(safetyplanId);
 
-        if(!copingStrategies.isEmpty())
-            copingStrategies.forEach(copingStrategy -> jpaCopingStrategyRepository.delete(copingStrategy));
+            if(!copingStrategies.isEmpty())
+                copingStrategies.forEach(copingStrategy -> jpaCopingStrategyRepository.delete(copingStrategy));
+        }
+        throw exception(SAFETYPLAN, ENTITY_NOT_FOUND, Long.toString(safetyplanId));
+    }
+
+    private RuntimeException exception(EntityType entityType, ExceptionType exceptionType, String... args) {
+        return CoveException.throwException(entityType, exceptionType, args);
     }
 }
